# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Machete.Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/TransportRules": {
      "get": {
        "tags": [
          "TransportRules"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransportRuleVM"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransportRuleVM"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransportRuleVM"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TransportRules"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              }
            }
          }
        }
      }
    },
    "/api/TransportRules/{id}": {
      "get": {
        "tags": [
          "TransportRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TransportRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransportRuleVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TransportRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportRuleVM"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransportCostRuleVM": {
        "type": "object",
        "properties": {
          "transportRuleId": {
            "type": "integer",
            "format": "int32"
          },
          "minWorker": {
            "type": "integer",
            "format": "int32"
          },
          "maxWorker": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "createdby": {
            "type": "string",
            "nullable": true
          },
          "datecreated": {
            "type": "string",
            "nullable": true
          },
          "dateupdated": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "updatedby": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransportRuleVM": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "lookupKey": {
            "type": "string",
            "nullable": true
          },
          "zoneLabel": {
            "type": "string",
            "nullable": true
          },
          "zipcodes": {
            "type": "string",
            "nullable": true
          },
          "costRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransportCostRuleVM"
            },
            "nullable": true
          },
          "createdby": {
            "type": "string",
            "nullable": true
          },
          "datecreated": {
            "type": "string",
            "nullable": true
          },
          "dateupdated": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "updatedby": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}